Few tweaks to content types as documentation changes + new view display.
I've changed the wording on the user stories child fields, as i didnt intend for them to sum AND add the current value, i've also changed the wording so its more consistant.
I also remvoed the 'strucutred order' field, as we can use draggable view's custom table instead and reduce complexity for the edit page.
The view should have the menu hooks for OA to work out of the box, but its not got full feature support yet.